FilterTransformer does not work as an exec plugin
There does not seem to be a way to remove resources using a transformer exec plugin (it is unclear if Go plugins have the ability; they would need direct access to the resource map).

This is likely by design as it would be difficult to implement it properly in the context of a kustomization.