Downgrade KYAML to v0.10.17
Starting in version v0.10.18 KYAML has a transitive dependency on version on gnostic v0.4.1 or better: in that version they changed the case of the package names as a "bug fix". Kubernetes code prior to 1.18 uses the old names, so we cannot share a dependency tree with 1.17 Kube API code unless we stop upgrading KYAML here.